use C11 _Generic macros for q_min, q_max and CLAMP, if available.
.. if not, then use GCC expression macros for them, if available.

+++ Quake/gl_model.c
@@ -508,7 +508,7 @@ void Mod_LoadTextures (lump_t *l)
 		if (((byte*)(mt+1) + pixels) > (mod_base + l->fileofs + l->filelen))
 		{
 			Con_DPrintf("Texture %s extends past end of lump\n", mt->name);
-			pixels = q_max(0, (mod_base + l->fileofs + l->filelen) - (byte*)(mt+1));
+			pixels = q_max(0L, (long)((mod_base + l->fileofs + l->filelen) - (byte*)(mt+1)));
 		}
 
 		tx->update_warp = false; //johnfitz
