+++ Quake/pl_osx.m
@@ -56,8 +56,8 @@ char *PL_GetClipboardData (void)
 	NSString* clipboardString = [pasteboard stringForType: NSPasteboardTypeString];
 	if (clipboardString != NULL && [clipboardString length] > 0) {
 		size_t sz = [clipboardString length] + 1;
-		sz = q_min(MAX_CLIPBOARDTXT, sz);
-		data = (char *) Z_Malloc(sz);
+		sz = q_min((size_t)(MAX_CLIPBOARDTXT), sz);
+		data = (char *) Z_Malloc((int)sz);
 #if (MAC_OS_X_VERSION_MIN_REQUIRED < 1040)	/* for ppc builds targeting 10.3 and older */
 		q_strlcpy (data, [clipboardString cString], sz);
 #else
